Add files via upload
Final uppload of Segmenting and clustering neighborhoods Toronto
Added to the Torontio ipynb-file is a couple of files with downloaded material i usein combination with the Toronto clustering file - due to the fact that it is always (if possible) beter to have the data stored on a local harddrive compared to always connect to Big Datasystems. Of course, sometimes one must have the freashest of data and then one must be connected to the cloud.